Steren ide patch

+++ scripts/packages/chocolatey/build.ps1
@@ -1,27 +1,54 @@
 param(
   [string] $version = "0.3.2",
-  [switch] $isRelease
+  [int] $rc = 0,
+  [switch] $fixPackage,
+  [string] $mode = "local",
+  [string] $checksum = ""
 )
 
-$tvVersion = $version
-$tvFilename = "bazel-$($version)-windows-x86_64.zip"
-if ($isRelease) {
-  $tvUri = "https://github.com/bazelbuild/bazel/releases/download/$($version)/$($tvFilename)"
-} else {
+$ErrorActionPreference = 'Stop'
+
+write-host "mode: $mode"
+if ($mode -eq "release") {
+  $tvVersion = $version
+  $tvFilename = "bazel-$($tvVersion)-windows-x86_64.zip"
+  $tvUri = "https://github.com/bazelbuild/bazel/releases/download/$($tvVersion)/$($tvFilename)"
+  $tvReleaseNotesUri = "https://github.com/bazelbuild/bazel/releases/tag/$tvVersion"
+} elseif ($mode -eq "rc") {
+  $tvVersion = "$($version)-rc$($rc)"
+  $tvFilename = "bazel-$($version)rc$($rc)-windows-x86_64.zip"
+  $tvUri = "https://releases.bazel.build/$($version)/rc$($rc)/$($tvFilename)"
+  $tvReleaseNotesUri = "https://releases.bazel.build/$($version)/rc$($rc)/index.html"
+} elseif ($mode -eq "local") {
+  $tvVersion = $version
+  $tvFilename = "bazel-$($tvVersion)-windows-x86_64.zip"
   $tvUri = "http://localhost:8000/$($tvFilename)"
+  $tvReleaseNotesUri = "http://localhost:8000/dummy"
+} else {
+  throw "mode parameter '$mode' unsupported. Please use local, rc, or release."
 }
-write-host "download uri: $($tvUri)"
 
-rm -force -ErrorAction SilentlyContinue ./*.nupkg
-rm -force -ErrorAction SilentlyContinue ./*.zip
-rm -force -ErrorAction SilentlyContinue ./bazel.nuspec
-rm -force -ErrorAction SilentlyContinue ./tools/chocolateyinstall.ps1
+if ($fixPackage -eq $true) {
+  $prefix = "-"
+  if ($mode -eq "release") {
+    $prefix = "."
+  }
+  $tvPackageFixVersion = "$($prefix)$((get-date).tostring("yyyyMMdd_hhmmss"))"
+}
+rm -force -ErrorAction SilentlyContinue "./*.nupkg"
+rm -force -ErrorAction SilentlyContinue "./bazel.nuspec"
+rm -force -ErrorAction SilentlyContinue "./tools/LICENSE*"
+rm -force -ErrorAction SilentlyContinue "./tools/*.orig"
+rm -force -ErrorAction SilentlyContinue "./tools/params.*"
+if ($checksum -eq "") {
+  rm -force -ErrorAction SilentlyContinue ./*.zip
+}
 
-if ($isRelease) {
+if (($mode -eq "release") -or ($mode -eq "rc")) {
   Invoke-WebRequest "$($tvUri).sha256" -UseBasicParsing -passthru -outfile sha256.txt
   $tvChecksum = (gc sha256.txt).split(' ')[0]
   rm sha256.txt
-} else {
+} elseif ($mode -eq "local") {
   Add-Type -A System.IO.Compression.FileSystem
   $outputDir = "$pwd/../../../output"
   $zipFile = "$pwd/$($tvFilename)"
@@ -34,24 +61,18 @@ $nuspecTemplate = get-content "bazel.nuspec.template" | out-string
 $nuspecExpanded = $ExecutionContext.InvokeCommand.ExpandString($nuspecTemplate)
 add-content -value $nuspecExpanded -path bazel.nuspec
 
-$installerScriptTemplate = get-content "chocolateyinstall.ps1.template" | out-string
-$installerScriptExpanded = $ExecutionContext.InvokeCommand.ExpandString($installerScriptTemplate)
-$installerScriptExpanded = $installerScriptExpanded -replace "ps_var_","$"
-$installerScriptExpanded = $installerScriptExpanded -replace "escape_char","``"
-add-content -value $installerScriptExpanded -path ./tools/chocolateyinstall.ps1
-
-$uninstallerScriptTemplate = get-content "chocolateyuninstall.ps1.template" | out-string
-$uninstallerScriptExpanded = $ExecutionContext.InvokeCommand.ExpandString($uninstallerScriptTemplate)
-$uninstallerScriptExpanded = $uninstallerScriptExpanded -replace "ps_var_","$"
-$uninstallerScriptExpanded = $uninstallerScriptExpanded -replace "escape_char","``"
-add-content -value $uninstallerScriptExpanded -path ./tools/chocolateyuninstall.ps1
-
-write-host "Copying LICENSE.txt from repo-root to tools directory"
+write-host "Copying LICENSE from repo-root to tools directory"
 $licenseHeader = @"
-From: https://github.com/bazelbuild/bazel/blob/master/LICENSE.txt
+From: https://github.com/bazelbuild/bazel/blob/master/LICENSE
 
 "@
-add-content -value $licenseHeader -path "./tools/LICENSE.txt"
-add-content -value (get-content "../../../LICENSE.txt") -path "./tools/LICENSE.txt"
+add-content -value $licenseHeader -path "./tools/LICENSE"
+add-content -value (get-content "../../../LICENSE") -path "./tools/LICENSE"
+
+$params = @"
+$tvUri
+$tvChecksum
+"@
+add-content -value $params -path "./tools/params.txt"
 
 choco pack ./bazel.nuspec
