+++ src/main/java/com/google/devtools/build/lib/runtime/commands/CleanCommand.java
@@ -20,7 +20,6 @@ import com.google.devtools.build.lib.buildtool.OutputDirectoryLinksUtils;
 import com.google.devtools.build.lib.events.Event;
 import com.google.devtools.build.lib.runtime.BlazeCommand;
 import com.google.devtools.build.lib.runtime.BlazeCommandDispatcher.ShutdownBlazeServerException;
-import com.google.devtools.build.lib.runtime.BlazeCommandDispatcher.ShutdownMethod;
 import com.google.devtools.build.lib.runtime.Command;
 import com.google.devtools.build.lib.runtime.CommandEnvironment;
 import com.google.devtools.build.lib.shell.CommandException;
@@ -32,92 +31,160 @@ import com.google.devtools.build.lib.util.ShellEscaper;
 import com.google.devtools.build.lib.vfs.FileSystemUtils;
 import com.google.devtools.build.lib.vfs.Path;
 import com.google.devtools.common.options.Option;
+import com.google.devtools.common.options.OptionDocumentationCategory;
+import com.google.devtools.common.options.OptionEffectTag;
 import com.google.devtools.common.options.OptionsBase;
 import com.google.devtools.common.options.OptionsParser;
 import com.google.devtools.common.options.OptionsProvider;
+import java.io.FileDescriptor;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.logging.LogManager;
 import java.util.logging.Logger;
 
-/**
- * Implements 'blaze clean'.
- */
-@Command(name = "clean",
-         builds = true,  // Does not, but people expect build options to be there
-         options = { CleanCommand.Options.class },
-         help = "resource:clean.txt",
-         shortDescription = "Removes output files and optionally stops the server.",
-         // TODO(bazel-team): Remove this - we inherit a huge number of unused options.
-         inherits = { BuildCommand.class })
+/** Implements 'blaze clean'. */
+@Command(
+  name = "clean",
+  builds = true, // Does not, but people expect build options to be there
+  writeCommandLog = false, // Do not create a command.log, otherwise we couldn't delete it.
+  options = {CleanCommand.Options.class},
+  help = "resource:clean.txt",
+  shortDescription = "Removes output files and optionally stops the server.",
+  // TODO(bazel-team): Remove this - we inherit a huge number of unused options.
+  inherits = {BuildCommand.class}
+)
 public final class CleanCommand implements BlazeCommand {
   /**
    * An interface for special options for the clean command.
    */
   public static class Options extends OptionsBase {
-    @Option(name = "clean_style",
-            defaultValue = "",
-            category = "clean",
-            help = "Can be either 'expunge' or 'expunge_async'.")
+    @Option(
+      name = "clean_style",
+      defaultValue = "",
+      category = "clean",
+      documentationCategory = OptionDocumentationCategory.UNCATEGORIZED,
+      effectTags = {OptionEffectTag.UNKNOWN},
+      help = "Can be 'expunge', 'expunge_async', or 'async'."
+    )
     public String cleanStyle;
 
-    @Option(name = "expunge",
-            defaultValue = "false",
-            category = "clean",
-            expansion = "--clean_style=expunge",
-            help = "If specified, clean will remove the entire working tree for this %{product} "
-                 + "instance, which includes all %{product}-created temporary and build output "
-                 + "files, and it will stop the %{product} server if it is running.")
-    public boolean expunge;
+    @Option(
+      name = "expunge",
+      defaultValue = "null",
+      category = "clean",
+      expansion = "--clean_style=expunge",
+      documentationCategory = OptionDocumentationCategory.UNCATEGORIZED,
+      effectTags = {OptionEffectTag.UNKNOWN},
+      help =
+          "If specified, clean removes the entire working tree for this %{product} "
+              + "instance, which includes all %{product}-created temporary and build output "
+              + "files, and stops the %{product} server if it is running."
+    )
+    public Void expunge;
 
-    @Option(name = "expunge_async",
-        defaultValue = "false",
-        category = "clean",
-        expansion = "--clean_style=expunge_async",
-        help = "If specified, clean will asynchronously remove the entire working tree for "
-             + "this %{product} instance, which includes all %{product}-created temporary and "
-             + "build output files, and it will stop the %{product} server if it is running. When "
-             + "this command completes, it will be safe to execute new commands in the same "
-             + "client, even though the deletion may continue in the background.")
-    public boolean expunge_async;
+    @Option(
+      name = "expunge_async",
+      defaultValue = "null",
+      category = "clean",
+      expansion = "--clean_style=expunge_async",
+      documentationCategory = OptionDocumentationCategory.UNCATEGORIZED,
+      effectTags = {OptionEffectTag.UNKNOWN},
+      help =
+          "If specified, clean asynchronously removes the entire working tree for "
+              + "this %{product} instance, which includes all %{product}-created temporary and "
+              + "build output files, and stops the %{product} server if it is running. When "
+              + "this command completes, it will be safe to execute new commands in the same "
+              + "client, even though the deletion may continue in the background."
+    )
+    public Void expungeAsync;
+
+    @Option(
+      name = "async",
+      defaultValue = "null",
+      category = "clean",
+      expansion = "--clean_style=async",
+      documentationCategory = OptionDocumentationCategory.UNCATEGORIZED,
+      effectTags = {OptionEffectTag.UNKNOWN},
+      help =
+          "If specified, clean asynchronously removes the entire working tree for "
+              + "this %{product} instance, which includes all %{product}-created temporary and "
+              + "build output files. When this command completes, it will be safe to execute new "
+              + "commands in the same client, even though the deletion may continue in the "
+              + "background."
+    )
+    public Void async;
   }
 
-  private static Logger LOG = Logger.getLogger(CleanCommand.class.getName());
+  /**
+   * Posted on the public event stream to announce that a clean is happening.
+   */
+  public static class CleanStartingEvent {
+    private final OptionsProvider optionsProvider;
+
+    public CleanStartingEvent(OptionsProvider optionsProvider) {
+      this.optionsProvider = optionsProvider;
+    }
+
+    public OptionsProvider getOptionsProvider() {
+      return optionsProvider;
+    }
+  }
+
+  private static final Logger logger = Logger.getLogger(CleanCommand.class.getName());
 
   @Override
   public ExitCode exec(CommandEnvironment env, OptionsProvider options)
       throws ShutdownBlazeServerException {
     Options cleanOptions = options.getOptions(Options.class);
-    cleanOptions.expunge_async = cleanOptions.cleanStyle.equals("expunge_async");
-    cleanOptions.expunge = cleanOptions.cleanStyle.equals("expunge");
+    boolean expungeAsync = cleanOptions.cleanStyle.equals("expunge_async");
+    boolean expunge = cleanOptions.cleanStyle.equals("expunge");
+    boolean async = cleanOptions.cleanStyle.equals("async");
 
     env.getEventBus().post(new NoBuildEvent());
 
-    if (!cleanOptions.expunge && !cleanOptions.expunge_async
-        && !cleanOptions.cleanStyle.isEmpty()) {
+    if (!expunge && !expungeAsync && !async && !cleanOptions.cleanStyle.isEmpty()) {
       env.getReporter().handle(Event.error(
           null, "Invalid clean_style value '" + cleanOptions.cleanStyle + "'"));
       return ExitCode.COMMAND_LINE_ERROR;
     }
 
+    String asyncName = (expunge || expungeAsync) ? "--expunge_async" : "--async";
+
     // TODO(dmarting): Deactivate expunge_async on non-Linux platform until we completely fix it
     // for non-Linux platforms (https://github.com/bazelbuild/bazel/issues/1906).
-    if (cleanOptions.expunge_async && OS.getCurrent() != OS.LINUX) {
-      env.getReporter().handle(Event.info(null /*location*/,
-          "--expunge_async cannot be used on non-Linux platforms, falling back to --expunge"));
-      cleanOptions.expunge_async = false;
-      cleanOptions.expunge = true;
-      cleanOptions.cleanStyle = "expunge";
+    // MacOS and FreeBSD support setsid(2) but don't have /usr/bin/setsid, so if we wanted to
+    // support --expunge_async on these platforms, we'd have to write a wrapper that calls setsid(2)
+    // and exec(2).
+    if ((expungeAsync || async) && OS.getCurrent() != OS.LINUX) {
+      String fallbackName = expungeAsync ? "--expunge" : "synchronous clean";
+      env.getReporter()
+          .handle(
+              Event.info(
+                  null /*location*/,
+                  asyncName
+                      + " cannot be used on non-Linux platforms, falling back to "
+                      + fallbackName));
+      expunge = expungeAsync;
+      expungeAsync = false;
+      async = false;
+      cleanOptions.cleanStyle = expunge ? "expunge" : "";
     }
 
-    String cleanBanner = cleanOptions.expunge_async ?
-        "Starting clean." :
-        "Starting clean (this may take a while). " +
-            "Consider using --expunge_async if the clean takes more than several minutes.";
+    String cleanBanner =
+        (expungeAsync || async)
+            ? "Starting clean."
+            : "Starting clean (this may take a while). "
+                + "Consider using "
+                + asyncName
+                + " if the clean takes more than several minutes.";
 
+    env.getEventBus().post(new CleanStartingEvent(options));
     env.getReporter().handle(Event.info(null/*location*/, cleanBanner));
+
     try {
       String symlinkPrefix = options.getOptions(BuildRequest.BuildRequestOptions.class)
           .getSymlinkPrefix(env.getRuntime().getProductName());
-      actuallyClean(env, env.getOutputBase(), cleanOptions, symlinkPrefix);
+      actuallyClean(env, env.getOutputBase(), expunge, expungeAsync, async, symlinkPrefix);
       return ExitCode.SUCCESS;
     } catch (IOException e) {
       env.getReporter().handle(Event.error(e.getMessage()));
@@ -131,14 +198,64 @@ public final class CleanCommand implements BlazeCommand {
     }
   }
 
-  private void actuallyClean(CommandEnvironment env,
-      Path outputBase, Options cleanOptions, String symlinkPrefix) throws IOException,
-      ShutdownBlazeServerException, CommandException, ExecException, InterruptedException {
+  private static void asyncClean(CommandEnvironment env, Path path, String pathItemName)
+      throws IOException, CommandException {
+    String tempBaseName = path.getBaseName() + "_tmp_" + ProcessUtils.getpid();
+
+    // Keeping tempOutputBase in the same directory ensures it remains in the
+    // same file system, and therefore the mv will be atomic and fast.
+    Path tempPath = path.getParentDirectory().getChild(tempBaseName);
+    path.renameTo(tempPath);
+    env.getReporter()
+        .handle(Event.info(null, pathItemName + " moved to " + tempPath + " for deletion"));
+
+    // Daemonize the shell and use the double-fork idiom to ensure that the shell
+    // exits even while the "rm -rf" command continues.
+    String command =
+        String.format(
+            "exec >&- 2>&- <&- && (/usr/bin/setsid /bin/rm -rf %s &)&",
+            ShellEscaper.escapeString(tempPath.getPathString()));
+
+    logger.info("Executing shell commmand " + ShellEscaper.escapeString(command));
+
+    // Doesn't throw iff command exited and was successful.
+    new CommandBuilder()
+        .addArg(command)
+        .useShell(true)
+        .setWorkingDir(tempPath.getParentDirectory())
+        .build()
+        .execute();
+  }
+
+  private void actuallyClean(
+      CommandEnvironment env,
+      Path outputBase,
+      boolean expunge,
+      boolean expungeAsync,
+      boolean async,
+      String symlinkPrefix)
+      throws IOException, ShutdownBlazeServerException, CommandException, ExecException,
+          InterruptedException {
+    String workspaceDirectory = env.getWorkspace().getBaseName();
     if (env.getOutputService() != null) {
       env.getOutputService().clean();
     }
-    if (cleanOptions.expunge) {
-      LOG.info("Expunging...");
+    env.getBlazeWorkspace().clearCaches();
+    if (expunge) {
+      logger.info("Expunging...");
+      env.getRuntime().prepareForAbruptShutdown();
+      // Close java.log.
+      LogManager.getLogManager().reset();
+      // Close the default stdout/stderr.
+      if (FileDescriptor.out.valid()) {
+        new FileOutputStream(FileDescriptor.out).close();
+      }
+      if (FileDescriptor.err.valid()) {
+        new FileOutputStream(FileDescriptor.err).close();
+      }
+      // Close the redirected stdout/stderr.
+      System.out.close();
+      System.err.close();
       // Delete the big subdirectories with the important content first--this
       // will take the most time. Then quickly delete the little locks, logs
       // and links right before we exit. Once the lock file is gone there will
@@ -146,52 +263,38 @@ public final class CleanCommand implements BlazeCommand {
       // all significant files will be gone by then.
       FileSystemUtils.deleteTreesBelow(outputBase);
       FileSystemUtils.deleteTree(outputBase);
-    } else if (cleanOptions.expunge_async) {
-      LOG.info("Expunging asynchronously...");
-      String tempBaseName = outputBase.getBaseName() + "_tmp_" + ProcessUtils.getpid();
-
-      // Keeping tempOutputBase in the same directory ensures it remains in the
-      // same file system, and therefore the mv will be atomic and fast.
-      Path tempOutputBase = outputBase.getParentDirectory().getChild(tempBaseName);
-      outputBase.renameTo(tempOutputBase);
-      env.getReporter().handle(Event.info(
-          null, "Output base moved to " + tempOutputBase + " for deletion"));
-
-      // Daemonize the shell and use the double-fork idiom to ensure that the shell
-      // exits even while the "rm -rf" command continues.
-      String command = String.format("exec >&- 2>&- <&- && (/usr/bin/setsid /bin/rm -rf %s &)&",
-          ShellEscaper.escapeString(tempOutputBase.getPathString()));
-
-      LOG.info("Executing shell commmand " + ShellEscaper.escapeString(command));
-
-      // Doesn't throw iff command exited and was successful.
-      new CommandBuilder().addArg(command).useShell(true)
-        .setWorkingDir(tempOutputBase.getParentDirectory())
-        .build().execute();
+    } else if (expungeAsync) {
+      logger.info("Expunging asynchronously...");
+      env.getRuntime().prepareForAbruptShutdown();
+      asyncClean(env, outputBase, "Output base");
     } else {
-      LOG.info("Output cleaning...");
-      env.getBlazeWorkspace().clearCaches();
+      logger.info("Output cleaning...");
+      env.getBlazeWorkspace().resetEvaluator();
       // In order to be sure that we delete everything, delete the workspace directory both for
       // --deep_execroot and for --nodeep_execroot.
-      for (String directory : new String[] {
-          env.getWorkspaceName(), "execroot/" + env.getWorkspaceName() }) {
+      for (String directory : new String[] {workspaceDirectory, "execroot"}) {
         Path child = outputBase.getRelative(directory);
         if (child.exists()) {
-          LOG.finest("Cleaning " + child);
-          FileSystemUtils.deleteTreesBelow(child);
+          logger.finest("Cleaning " + child + (async ? " asynchronously..." : ""));
+          if (async) {
+            asyncClean(env, child, "Output tree");
+          } else {
+            FileSystemUtils.deleteTreesBelow(child);
+          }
         }
       }
     }
     // remove convenience links
     OutputDirectoryLinksUtils.removeOutputDirectoryLinks(
-        env.getWorkspaceName(), env.getWorkspace(), env.getReporter(), symlinkPrefix,
-        env.getRuntime().getProductName());
+        workspaceDirectory, env.getWorkspace(), env.getReporter(),
+        symlinkPrefix, env.getRuntime().getProductName());
     // shutdown on expunge cleans
-    if (cleanOptions.expunge || cleanOptions.expunge_async) {
-      throw new ShutdownBlazeServerException(0, ShutdownMethod.EXPUNGE);
+    if (expunge || expungeAsync) {
+      throw new ShutdownBlazeServerException(0);
     }
+    System.gc();
   }
 
   @Override
-  public void editOptions(CommandEnvironment env, OptionsParser optionsParser) {}
+  public void editOptions(OptionsParser optionsParser) {}
 }
